requirements.txt
beautifulsoup4==4.5.3
bs4==0.0.1
click==6.7
Flask==0.12
future==0.16.0
html5lib==0.999999999
itsdangerous==0.24
Jinja2==2.9.5
line-bot-sdk==1.1.0
lxml==3.7.3
MarkupSafe==1.0
numpy==1.12.0
pandas==0.19.2
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
six==1.10.0
webencodings==0.5
Werkzeug==0.12
gunicorn==19.6.0
